backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
pydantic-settings==2.1.0
supabase==2.28.0
pyhanko==0.21.0
pyhanko-certvalidator==0.26.3
cryptography==41.0.7
python-dotenv==1.0.0
httpx==0.28.1
aiofiles==23.2.1
reportlab==4.2.5
pypdf==5.1.0
